implemented enrollment accept/reject and checkin/cancel

+++ src/main/java/com/tistory/jaimemin/studyrecruitment/event/EventService.java
@@ -63,8 +63,29 @@ public class EventService {
 
     public void cancelEnrollment(Event event, Account account) {
         Enrollment enrollment = enrollmentRepository.findByEventAndAccount(event, account);
+
+        if (enrollment.isAttended()) {
+            return;
+        }
+
         event.removeEnrollment(enrollment);
         enrollmentRepository.delete(enrollment);
         event.acceptNextWaitingEnrollment();
+    }
+
+    public void acceptEnrollment(Event event, Enrollment enrollment) {
+        event.accept(enrollment);
+    }
+
+    public void rejectEnrollment(Event event, Enrollment enrollment) {
+        event.reject(enrollment);
+    }
+
+    public void checkInEnrollment(Enrollment enrollment) {
+        enrollment.setAttended(true);
+    }
+
+    public void cancelCheckInEnrollment(Enrollment enrollment) {
+        enrollment.setAttended(false);
     }
 }
